pyproject.toml
[tool.poetry]
name = "definitive-guide-to-dimensional-modeling"
version = "0.1.0"
description = ""
authors = ["sillians <ihuomacbasil@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
black = "^24.8.0"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
pylint = "^3.2.6"
setuptools = "^74.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
